Some work on changing comments, log messages, and variable and function names to reflect that the plugins are not plugins anymore.
git-svn-id: https://dolphin-emu.googlecode.com/svn/trunk@7170 8ced0084-cf51-0410-be5f-012b33b47a6e

+++ Source/Core/VideoCommon/Src/VideoConfig.cpp
@@ -222,7 +222,7 @@ void VideoConfig::GameIniLoad(const char *ini_file)
 void VideoConfig::VerifyValidity()
 {
 	// TODO: Check iMaxAnisotropy value
-	if (iAdapter >= backend_info.Adapters.size()) iAdapter = 0;
+	if (iAdapter >= (int)backend_info.Adapters.size()) iAdapter = 0;
 	if (!backend_info.bSupportsEFBToRAM) bCopyEFBToTexture = true;
 	if (iMultisampleMode < 0 || iMultisampleMode >= (int)backend_info.AAModes.size()) iMultisampleMode = 0;
 	if (!backend_info.bSupportsRealXFB) bUseRealXFB = false;
